Remove dependency of java.math.BigInteger

+++ src/jvmMain/kotlin/org/bisdk/sdk/BiPackage.kt
@@ -1,7 +1,6 @@
 package org.bisdk.sdk
 
 import org.bisdk.*
-import java.math.BigInteger
 
 /**
  * A BiPackage is sent over the socket to the gateway. It contains the command and the payload.
@@ -67,7 +66,7 @@ data class BiPackage(
             idx += Lengths.TOKEN_BYTES
             var commandInt = (ba.copyOfRange(idx, idx + Lengths.COMMAND_BYTES).toHexString()).toInt(16)
             var isResponse = false
-            if(BigInteger.valueOf(commandInt.toLong()).testBit(7)) {
+            if(commandInt.testBit(7)) {
                commandInt  = commandInt xor (1 shl 7)
                 isResponse = true
             }
